Ensure each module has its own unique Android namespace
