+++ 119602-consultation/build.gradle.kts
@@ -124,7 +124,7 @@ kotlin {
 
 // Android configuration
 android {
-    namespace = properties["namespace"].toString()
+    namespace = "eu.europa.ec.eudi.etsi119602.consultation"
     group = properties["group"].toString()
     compileSdk = properties["android.targetSdk"].toString().toInt()
