+++ 119602-consultation/src/commonMain/kotlin/eu/europa/ec/eudi/etsi119602/consultation/eu/EUPubEAAProvidersList.kt
@@ -15,10 +15,10 @@
  */
 package eu.europa.ec.eudi.etsi119602.consultation.eu
 
-import eu.europa.ec.eudi.etsi119602.CountryCode
-import eu.europa.ec.eudi.etsi119602.ETSI19602
-import eu.europa.ec.eudi.etsi119602.MultiLanguageURI
-import eu.europa.ec.eudi.etsi119602.Uri
+import eu.europa.ec.eudi.etsi119602.datamodel.CountryCode
+import eu.europa.ec.eudi.etsi119602.datamodel.ETSI19602
+import eu.europa.ec.eudi.etsi119602.datamodel.MultiLanguageURI
+import eu.europa.ec.eudi.etsi119602.datamodel.Uri
 
 public val EUPubEAAProvidersList: EUListOfTrustedEntitiesProfile =
     EUListOfTrustedEntitiesProfile(
