+++ 119602-data-model/build.gradle.kts
@@ -100,7 +100,7 @@ kotlin {
 
 // Android configuration
 android {
-    namespace = properties["namespace"].toString()
+    namespace = "eu.europa.ec.eudi.etsi119602.datamodel"
     group = properties["group"].toString()
     compileSdk = properties["android.targetSdk"].toString().toInt()
