--- 119602-data-model/src/commonMain/kotlin/eu/europa/ec/eudi/etsi119602/Uri.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2026 European Commission
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eu.europa.ec.eudi.etsi119602
-
-import kotlinx.serialization.Serializable
-import kotlin.jvm.JvmInline
-
-@Serializable
-@JvmInline
-public value class Uri(public val value: String) {
-
-    override fun toString(): String = value
-
-    public companion object {
-        public fun parse(value: String): Uri {
-            require(value.isNotBlank()) { "Uri cannot be blank" }
-            return Uri(value)
-        }
-    }
-}
