+++ 119602-data-model/src/commonMain/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/ETSI19602.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
 /**
  * [ETSI 119 602](https://www.etsi.org/deliver/etsi_ts/119600_119699/119602/01.01.01_60/ts_119602v010101p.pdf)
