+++ 119602-data-model/src/commonMain/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/Language.kt
@@ -13,34 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
 import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
 
 @Serializable
 @JvmInline
-public value class CountryCode
+public value class Language
 @Throws(IllegalArgumentException::class)
-public constructor(
-    public val value: String,
-) {
+public constructor(public val value: String) {
     init {
-        require(value.isNotBlank()) { "Country code cannot be blank" }
-        require(value.matches(CAPITAL_LETTERS_PATTERN)) { "Country code must be all capital letters" }
+        Assertions.requireNotBlank(value, ETSI19602.LANG)
+        require(value.matches(ALPHA_2_PATTERN)) { "Invalid ${ETSI19602.LANG}" }
     }
 
     override fun toString(): String = value
 
     public companion object {
-        public val EU: CountryCode get() = iso3166(ETSI19602.COUNTRY_CODE_EU)
-        private val CAPITAL_LETTERS_PATTERN = Regex("^[A-Z]+$")
-        private val ISO3166_1_ALPHA_2_PATTERN = Regex("^[A-Z]{2}$")
-
-        @Throws(IllegalArgumentException::class)
-        public fun iso3166(value: String): CountryCode {
-            require(value.matches(ISO3166_1_ALPHA_2_PATTERN)) { "Invalid ISO 3166-1 alpha-2 code: $value" }
-            return CountryCode(value)
-        }
+        public val ENGLISH: Language get() = Language(ETSI19602.LANG_ENGLISH)
+        private val ALPHA_2_PATTERN = Regex("^[a-z]{2}$")
     }
 }
