+++ 119602-data-model/src/commonMain/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/ListAndSchemeInformation.kt
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
-import eu.europa.ec.eudi.etsi119602.ListAndSchemeInformation.Companion.explicit
-import eu.europa.ec.eudi.etsi119602.ListAndSchemeInformation.Companion.implicit
+import eu.europa.ec.eudi.etsi119602.datamodel.ListAndSchemeInformation.Companion.explicit
+import eu.europa.ec.eudi.etsi119602.datamodel.ListAndSchemeInformation.Companion.implicit
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.Required
 import kotlinx.serialization.SerialName
