+++ 119602-data-model/src/jvmAndAndroidMain/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/ListOfTrustedEntities.jvmAndAndroid.kr.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
 import java.security.Provider
 import java.security.cert.X509Certificate
