+++ 119602-data-model/src/jvmAndAndroidMain/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/ServiceDigitalIdentity.jvmAndAndroid.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
 import java.security.NoSuchProviderException
 import java.security.Provider
