+++ 119602-data-model/src/jvmAndAndroidTest/kotlin/eu/europa/ec/eudi/etsi119602/datamodel/PKIObjectTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.etsi119602
+package eu.europa.ec.eudi.etsi119602.datamodel
 
 import kotlinx.serialization.json.Json
 import kotlin.test.Test
