+++ consultation-dss/build.gradle.kts
@@ -131,7 +131,7 @@ kotlin {
 
 // Android configuration
 android {
-    namespace = properties["namespace"].toString()
+    namespace = "eu.europa.ec.eudi.etsi1196x2.consultation.dss"
     group = properties["group"].toString()
     compileSdk = properties["android.targetSdk"].toString().toInt()
