+++ gradle.properties
@@ -20,7 +20,6 @@ android.minSdk=26
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 
 # Publishing settings
-namespace=eu.europa.ec.eudi.etsi1196x2
 group=eu.europa.ec.eudi
 version=0.4.0-alpha.1-SNAPSHOT
